some refactor about code comments (#20821)
